Move compressed-probe dispatch into JoinHashTable::Probe
