+++ src/execution/operator/join/physical_hash_join.cpp
@@ -3,7 +3,6 @@
 #include "duckdb/common/assert.hpp"
 #include "duckdb/common/projection_index.hpp"
 #include "duckdb/common/radix_partitioning.hpp"
-#include "duckdb/common/vector/dictionary_vector.hpp"
 #include "duckdb/common/types.hpp"
 #include "duckdb/common/types/value_map.hpp"
 #include "duckdb/common/uhugeint.hpp"
@@ -1576,7 +1575,9 @@ SinkFinalizeType PhysicalHashJoin::Finalize(Pipeline &pipeline, Event &event, Cl
 //===--------------------------------------------------------------------===//
 // Operator
 //===--------------------------------------------------------------------===//
-//! Structural gate for the compressed-vector probe paths; evaluated once per operator state
+//! Structural gate for the compressed-vector probe paths; evaluated once per operator state.
+//! Decides whether to allocate ProbeState::dict_state; Probe() dispatches into the fast paths
+//! whenever that allocation is present.
 static bool CanUseCompressedProbe(const HashJoinGlobalSinkState &sink, const vector<JoinCondition> &conditions) {
 	if (sink.external) {
 		// external joins re-finalize the HT mid-probe, invalidating the per-id pointer cache
@@ -1601,19 +1602,6 @@ static bool CanUseCompressedProbe(const HashJoinGlobalSinkState &sink, const vec
 	return true;
 }
 
-//! Per-chunk fast-reject so non-dictionary chunks skip the TryProbeDictionary call entirely
-static bool LHSChunkIsDictionaryEligible(const Vector &lhs_key) {
-	if (lhs_key.GetVectorType() != VectorType::DICTIONARY_VECTOR) {
-		return false;
-	}
-	return DictionaryVector::DictionarySize(lhs_key).IsValid();
-}
-
-//! Per-chunk fast-reject so non-constant chunks skip the TryProbeConstant call entirely
-static bool LHSChunkIsConstant(const Vector &lhs_key) {
-	return lhs_key.GetVectorType() == VectorType::CONSTANT_VECTOR;
-}
-
 class HashJoinOperatorState : public CachingOperatorState {
 public:
 	HashJoinOperatorState(ClientContext &context, const PhysicalHashJoin &op_p, HashJoinGlobalSinkState &sink)
@@ -1633,8 +1621,6 @@ public:
 	JoinHashTable::ProbeState probe_state;
 	//! Chunk to sink data into for external join
 	DataChunk spill_chunk;
-	//! Cached result of CanUseCompressedProbe for this operator state
-	bool compressed_probe_enabled = false;
 
 public:
 	void Finalize(const PhysicalOperator &op, ExecutionContext &context) override {
@@ -1688,7 +1674,10 @@ unique_ptr<OperatorState> PhysicalHashJoin::GetOperatorState(ExecutionContext &c
 		sink.InitializeProbeSpill();
 	}
 
-	state->compressed_probe_enabled = CanUseCompressedProbe(sink, conditions);
+	if (CanUseCompressedProbe(sink, conditions)) {
+		// non-null dict_state is the enabling signal for the compressed-vector probe paths in Probe()
+		state->probe_state.dict_state = make_uniq<JoinHashTable::ProbeDictionaryState>();
+	}
 
 	return std::move(state);
 }
@@ -1741,14 +1730,6 @@ OperatorResultType PhysicalHashJoin::ExecuteInternal(ExecutionContext &context, 
 			sink.hash_table->ProbeAndSpill(state.scan_structure, state.lhs_join_keys, state.join_key_state,
 			                               state.probe_state, input, *sink.probe_spill, state.spill_state,
 			                               state.spill_chunk);
-		} else if (state.compressed_probe_enabled && LHSChunkIsConstant(state.lhs_join_keys.data[0]) &&
-		           sink.hash_table->TryProbeConstant(state.scan_structure, state.lhs_join_keys, state.join_key_state,
-		                                             state.probe_state)) {
-			// scan_structure populated by the constant-vector fast path
-		} else if (state.compressed_probe_enabled && LHSChunkIsDictionaryEligible(state.lhs_join_keys.data[0]) &&
-		           sink.hash_table->TryProbeDictionary(state.scan_structure, state.lhs_join_keys, state.join_key_state,
-		                                               state.probe_state)) {
-			// scan_structure populated by the dictionary-aware path
 		} else {
 			sink.hash_table->Probe(state.scan_structure, state.lhs_join_keys, state.join_key_state, state.probe_state);
 		}
